Prevent huge step in initial smoothing
Due to the first iteration of the smoothing being called with huge
dt, there was a huge step in the smoothed waypoint.

This commit changes the initial values for the time stamps to large values,
such that the first iteration has a negative dt.

In addition, the smoothing checks for dt which is too small.